pass ebc as a float

+++ src/beer/malt.rs
@@ -3,22 +3,22 @@ use serde::{Deserialize, Serialize};
 #[derive(Serialize, Deserialize, Debug)]
 pub struct Malt {
     pub quantity: i32,
-    pub ebc: i32,
+    pub ebc: f32,
 }
 
 #[derive(Serialize, Debug)]
 pub struct Color(i32, i32, i32);
 
-pub fn calculate_ebc(efficiency: i32, quantity: i32, malts: &Vec<Malt>) -> i32 {
+pub fn calculate_ebc(efficiency: i32, quantity: i32, malts: &Vec<Malt>) -> f32 {
     let mut sum = 0.0;
     for malt in malts {
-        sum = sum + ((malt.quantity as f32 / 1000 as f32) * malt.ebc as f32);
+        sum = sum + ((malt.quantity as f32 / 1000 as f32) * malt.ebc);
     }
-    ((sum * 10.0 * (efficiency as f32 / 100.0)) / quantity as f32) as i32
+    (sum * 10.0 * (efficiency as f32 / 100.0)) / quantity as f32
 }
 
-pub fn calculate_color(ebc: i32) -> Color {
-    match ebc {
+pub fn calculate_color(ebc: f32) -> Color {
+    match ebc as i32 {
         1 => Color(255, 230, 153),
         2 => Color(255, 216, 120),
         3 => Color(255, 202, 90),
@@ -69,30 +69,30 @@ mod tests {
 
     #[test]
     fn test_calculate_ebc() {
-        assert_eq!(0, calculate_ebc(80, 20, &Vec::new()));
+        assert_eq!(0.0, calculate_ebc(80, 20, &Vec::new()));
         assert_eq!(
-            60,
+            60.0,
             calculate_ebc(
                 80,
                 20,
                 &vec![Malt {
-                    ebc: 150,
+                    ebc: 150.0,
                     quantity: 1000
                 }]
             )
         );
         assert_eq!(
-            64,
+            64.0,
             calculate_ebc(
                 80,
                 20,
                 &vec![
                     Malt {
-                        ebc: 150,
+                        ebc: 150.0,
                         quantity: 1000
                     },
                     Malt {
-                        ebc: 20,
+                        ebc: 20.0,
                         quantity: 500
                     },
                 ]
@@ -102,17 +102,17 @@ mod tests {
 
     #[test]
     fn test_calculate_color() {
-        let mut color = calculate_color(1);
+        let mut color = calculate_color(1.0);
         assert_eq!(255, color.0);
         assert_eq!(230, color.1);
         assert_eq!(153, color.2);
 
-        color = calculate_color(25);
+        color = calculate_color(25.0);
         assert_eq!(123, color.0);
         assert_eq!(26, color.1);
         assert_eq!(0, color.2);
 
-        color = calculate_color(70);
+        color = calculate_color(70.0);
         assert_eq!(0, color.0);
         assert_eq!(0, color.1);
         assert_eq!(0, color.2);
